feat: separate ADD form & refactor group-by/filters list
