+++ apps/web/src/common/modules/widgets/_constants/data-table-constant.ts
@@ -1,4 +1,6 @@
 
+import { GROUP_BY } from '@/services/cost-explorer/constants/cost-explorer-constant';
+
 
 export const DATA_TABLE_TYPE = {
     ADD: 'ADD',
@@ -10,3 +12,6 @@ export const DATA_SOURCE_DOMAIN = {
     ASSET: 'ASSET',
     SECURITY: 'SECURITY',
 };
+
+// NOTE: temporary solution for the global variable
+export const MANAGED_GLOBAL_VARIALBE = [GROUP_BY.WORKSPACE, GROUP_BY.PROJECT, GROUP_BY.SERVICE_ACCOUNT, GROUP_BY.REGION] as string[];
